add cog checks
